Reviewed-on: https://gitea.patrykkazmierczak.org/kazik/TCPManager/pulls/28

+++ src/main/java/org/tcpmanager/tcpmanager/statistics/intakehistory/IntakeHistory.java
@@ -30,7 +30,7 @@ public class IntakeHistory {
   private Date date;
 
   @Column(nullable = false)
-  private BigDecimal calories;
+  private Integer calories;
 
   @Column(nullable = false)
   private BigDecimal protein;
