+++ src/main/java/org/tcpmanager/tcpmanager/statistics/intakehistory/dto/IntakeHistoryResponse.java
@@ -3,7 +3,7 @@ package org.tcpmanager.tcpmanager.statistics.intakehistory.dto;
 import java.math.BigDecimal;
 import java.sql.Date;
 
-public record IntakeHistoryResponse(Long id, Date date, BigDecimal calories, BigDecimal protein,
+public record IntakeHistoryResponse(Long id, Date date, Integer calories, BigDecimal protein,
                                     BigDecimal fat, BigDecimal carbs, Integer caloriesGoal,
                                     Integer proteinGoal, Integer fatGoal, Integer carbsGoal,
                                     String username) {
