+++ src/main/java/org/tcpmanager/tcpmanager/user/events/MealAddedEvent.java
@@ -5,7 +5,7 @@ import java.sql.Date;
 import org.springframework.modulith.NamedInterface;
 
 @NamedInterface
-public record MealAddedEvent(Date date, String username, BigDecimal calories, BigDecimal protein,
+public record MealAddedEvent(Date date, String username, Integer calories, BigDecimal protein,
                              BigDecimal fat, BigDecimal carbs) {
 
 }
